refactor(type-declaration): separated type declaration into different files and folders and export them from a central index.ts file (barrel file) for better organization
