+++ src/providers/TodoProvider.tsx
@@ -1,7 +1,7 @@
 import { TodoContext } from '@/context';
 import { db } from '@/db/db';
 import { useAuth } from '@/hooks/useAuth';
-import type { DecryptedTodoType } from '@/types/types';
+import type { DecryptedTodoType } from '@/types';
 import { decryptString } from '@/utils/dcryptString';
 import { encryptString } from '@/utils/encryptString';
 import { useLiveQuery } from 'dexie-react-hooks';
