+++ src/types/api-types/authResponseType.ts
@@ -1,0 +1,6 @@
+import type { UserType } from '../model-types/userType';
+
+export type AuthResponseType = {
+	token: string;
+	user: UserType;
+};
