+++ src/types/api-types/categoryResponseType.ts
@@ -1,0 +1,5 @@
+import type { CategoryType } from '../model-types/categoryType';
+
+export type CategoriesResponse = {
+	categories: CategoryType[];
+};
