+++ src/types/api-types/todoResponseType.ts
@@ -1,0 +1,5 @@
+import type { TodoType } from '../model-types/todoType';
+
+export type TodoResponseType = {
+	todos: TodoType[];
+};
