+++ src/types/component-types/deleteModalType.ts
@@ -1,0 +1,6 @@
+export type DeleteModalProps = {
+	open: boolean;
+	onClose: () => void;
+	onConfirm: () => void;
+	itemName: string;
+};
