+++ src/types/component-types/editModalTypes.ts
@@ -1,0 +1,8 @@
+import type { TodoType } from '../model-types/todoType';
+
+export type EditModalProps = {
+	open: boolean;
+	onClose: () => void;
+	todo: TodoType | null;
+	onConfirm: (updatedTodo: TodoType) => void;
+};
