+++ src/types/erro-type/errorType.ts
@@ -1,0 +1,3 @@
+export type ErrorType = {
+	error: string;
+};
