+++ src/types/index.ts
@@ -1,0 +1,25 @@
+// context types
+export * from './context-types/authContextType';
+export * from './context-types/todoContextType';
+
+// components
+export * from './component-types/deleteModalType';
+export * from './component-types/editModalTypes';
+
+// models
+export * from './model-types/categoryType';
+export * from './model-types/decryptedType';
+export * from './model-types/sharedListType';
+export * from './model-types/todoType';
+export * from './model-types/userType';
+
+// api response
+export * from './api-types/authResponseType';
+export * from './api-types/categoryResponseType';
+export * from './api-types/todoResponseType';
+
+// pagination
+export * from './pagination-types/paginationButtonsComponentTypes';
+
+// error
+export * from './erro-type/errorType';
