+++ src/types/model-types/categoryType.ts
@@ -1,0 +1,7 @@
+export type CategoryType = {
+	id: string;
+	name: string;
+	color: string;
+	userId: string;
+	createdAt: string;
+};
