+++ src/types/model-types/decryptedType.ts
@@ -1,0 +1,8 @@
+export type DecryptedTodoType = {
+	id?: number;
+	userId?: number;
+	task: string;
+	status: 'pending' | 'completed';
+	deadline: string;
+	createdAt: string;
+};
