+++ src/types/model-types/sharedListType.ts
@@ -1,0 +1,11 @@
+export type SharedList = {
+	id: string;
+	name: string;
+	description: string;
+	ownerId: string;
+	createdAt: string;
+	updatedAt: string;
+	role: string;
+	status: string;
+	ownerName: string;
+};
