+++ src/types/model-types/userType.ts
@@ -1,0 +1,10 @@
+export type UserType = {
+	id: string;
+	email: string;
+	username: string;
+	firstName: string;
+	lastName: string;
+	avatarUrl: string;
+	createdAt: string;
+	updatedAt: string;
+};
